% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_annotation.R
\name{asr_annotation}
\alias{asr_annotation}
\title{converting result mapping to ASDB}
\usage{
asr_annotation(
  converter.result,
  gene.model = "Ensembl",
  genome.version = "hg38",
  gsea.gene.list = "",
  result.dir = ""
)
}
\arguments{
\item{converter.result}{asr_converter(rMATS_converter, SUPPA_converter, or spliceR_converter) result}

\item{gene.model}{gene model of reference. One of Refseq, Ensembl, or GENCODE.}

\item{genome.version}{genome version of reference. One of hg18, GRCh19, or GRCh38.}

\item{gsea.gene.list}{optional. reference gene list for gene enrichment test with annotation result gene list and knowledge-based database gene list. If gene list is empty, use all genes in reference.}

\item{result.dir}{directory where annotation result(.tsv file) is saved}
}
\value{
annotation result
}
\description{
converting result mapping to ASDB
}
\examples{
annotation.result.dir <- system.file(“extdata/annotation_result”, package=“ASpediaR”)
annotation.result <- asr_annotation(rmats.converter.result, “Ensembl”, “GRCh38”,
                                      result.dir=annotation.result.dir)
}
