% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_plot.R, R/asr_plot_ppi.R
\name{asr_plot}
\alias{asr_plot}
\title{visualization of annotation result}
\usage{
asr_plot(
  annotation.result,
  gtf.file.name,
  gene.model = "Ensembl",
  genome.version = "hg38",
  gene.name = "",
  as.id = "",
  heights.list = "",
  plot.data.list = ""
)

asr_plot(
  annotation.result,
  gtf.file.name,
  gene.model = "Ensembl",
  genome.version = "hg38",
  gene.name = "",
  as.id = "",
  heights.list = "",
  plot.data.list = ""
)
}
\arguments{
\item{annotation.result}{asr_annotation function result.}

\item{gtf.file.name}{a GTF format file of reference.}

\item{gene.model}{gene model of reference.}

\item{genome.version}{genome version of reference.}

\item{gene.name}{gene name to be visualization.}

\item{as.id}{list of AS ID to be visualization.}

\item{heights.list}{positive integer vectors for track heights include height of ideogram and gene track.}

\item{plot.data.list}{list of DNA or RNA feature to be visualization.choose from “conservation”, “NMD”, “repeats”, “domain”, “PTM”, and “RBP”.}
}
\description{
visualization of annotation result

visualization of annotation result
}
\examples{
gtf.file.name <- “examples/test_gtf.gtf”
asr_plot(annotation.result, gtf.file.name, gene.model="Ensembl", genome.version="hg38",
          gene.name="FGFR2")
asr_plot(annotation.result, gtf.file.name, gene.model="Ensembl", genome.version="hg38",
          gene.name="FGFR2", heights.list=c(1, 2, 2, 1, 3),
          list.of.plot=c(“conservation”, “NMD”, “RBP”))
asr_plot(annotation.result, gtf.file.name, gene.model="Ensembl", genome.version="hg38",
          as.id="chr10:121520169:121519979:121518829:121518682:121517463:121517319")
gtf.file.name <- “examples/test_gtf.gtf”
asr_plot(annotation.result, gtf.file.name, gene.model="Ensembl", genome.version="hg38",
          gene.name="FGFR2")
asr_plot(annotation.result, gtf.file.name, gene.model="Ensembl", genome.version="hg38",
          gene.name="FGFR2", heights.list=c(1, 2, 2, 1, 3),
          list.of.plot=c(“conservation”, “NMD”, “RBP”))
asr_plot(annotation.result, gtf.file.name, gene.model="Ensembl", genome.version="hg38",
          as.id="chr10:121520169:121519979:121518829:121518682:121517463:121517319")
}
