% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_plot_ppi.R
\name{asr_plot_ppi}
\alias{asr_plot_ppi}
\title{visualization of PPI network}
\usage{
asr_plot_ppi(annotation.result, gene.name = "", as.id = "", result.dir = "")
}
\arguments{
\item{annotation.result}{asr_annotation function result.}

\item{gene.name}{gene name to be visualization.}

\item{as.id}{list of AS ID to be visualization.}

\item{result.dir}{directory where PPI plots(.png files) are saved}
}
\description{
visualization of PPI network
}
\examples{
ppi.result.dir <- system.file(“extdata/ppi_result”, package=“ASpediaR”)

##using gene name
asr_plot_ppi(annotation.result, gene.name="FGFR2", result.dir=ppi.result.dir)

##using AS ID
asr_plot_ppi(annotation.result, 
            as.id="chr10:121520169:121519979:121518829:121518682:121517463:121517319",
            result.dir=ppi.result.dir)
}
