% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining_gsea.R
\name{mining_gsea}
\alias{mining_gsea}
\title{gene enrichment test between annotation gene list and knowledge-based database gene list.}
\usage{
mining_gsea(annotation.gene.list, whole.gene.list, result.dir)
}
\arguments{
\item{annotation.gene.list}{gene list from asr_annotation function result}

\item{whole.gene.list}{gene list from reference}
}
\description{
gene enrichment test between annotation gene list and knowledge-based database gene list.
}
\examples{
library(rtracklayer)
gtf.file.name <- “examples/test_gtf.gtf”
gtf.data <- import(gtf.file.name)
reference.gene.list <- unique(gtf.data$gene_name)
annotation.gene.list <- unique(annotation.result$gene_symbol)
mining_gsea(annotation.gene.list, reference.gene.list)
}
