% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMATS_converter.R
\name{rMATS.converter}
\alias{rMATS.converter}
\title{Generate ASpedia-R input format from rMATS result.}
\usage{
rMATS.converter(
  rMATS.result = "",
  as.type = "",
  pvalue.cutoff = 0.05,
  dpsi.cutoff = 0.1
)
}
\arguments{
\item{rMATS.result}{name of rMATS result file.}

\item{as.type}{AS event type. One of A3SS, A5SS, SE, MXE, or RI.}

\item{pvalue.cutoff}{value of pvalue cutoff. default value is 0.05}

\item{dpsi.cutoff}{value of dPSI cutoff. default value is 0.1}
}
\value{
converting.result
}
\description{
rMSTS result convert to ASpedia-R input format.
}
\examples{
rmats.result.file.name <- system.file(“extdata”, “rMATS_test.txt”, package=“ASpediaR”)
rmats.converter.result <- rMATS_converter(rmats.result.file.name, program=“rMATS”,
                                            as.type=“SE”)
}
