% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceR_converter.R
\name{spliceR.converter}
\alias{spliceR.converter}
\title{spliceR result convert to ASpedia-R input format.}
\usage{
spliceR.converter(spliceR.result, pvalue.cutoff, gene.model, genome.version)
}
\arguments{
\item{spliceR.result}{name of spliceR result file.}

\item{pvalue.cutoff}{value of pvalue cutoff. default value is 0.05}

\item{gene.model}{gene model of reference.}

\item{genome.version}{genome version of reference.}
}
\value{
converting.result
}
\description{
spliceR result convert to ASpedia-R input format.
}
\examples{
spliceR.result.file <- “examples/spliceR_test.txt”
splilcer.converter.result <- spliceR_converter(spliceR.result.file, program=”spliceR”,
                                         gene.model=”Ensembl’”, genome.version=”GRCh38”)
}
