% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUPPA_converter.R
\name{SUPPA.converter}
\alias{SUPPA.converter}
\title{Generate ASpedia-R input format from SUPPA result.}
\usage{
SUPPA.converter(
  SUPPA.result,
  pvalue.cutoff,
  dpsi.cutoff,
  gtf.file.name,
  ioe.file.name
)
}
\arguments{
\item{SUPPA.result}{name of SUPPA result file.}

\item{pvalue.cutoff}{value of pvalue cutoff. default value is 0.05}

\item{dpsi.cutoff}{value of dPSI cutoff. default value is 0.1}

\item{gtf.file.name}{a GTF format file of reference.}

\item{ioe.file.name}{name of ioe file generated by SUPPA generateEvents command.}
}
\value{
converting.result
}
\description{
Generate ASpedia-R input format from SUPPA result.
}
\examples{
suppa.result.file <- “examples/SUPPA_test.txt”
gtf.file.name <- “examples/test_gtf.gtf”
suppa.ioe.file.name <- “examples/SUPPA_test.ioe”
suppa.converter.result <- SUPPA_converter(suppa.result.file, gtf.file=gtf.file.name, ioe.file=suppa.ioe.file.name)
}
