% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_converter.R
\name{asr_converter}
\alias{asr_converter}
\title{Generate ASpedia-R input format from DAS analysis tools result.}
\usage{
asr_converter(
  das.analysis.result = "",
  program = "",
  pvalue.cutoff = 0.05,
  dpsi.cutoff = 0.1,
  gene.model = "Ensembl",
  genome.version = "hg38",
  as.type = "",
  gtf.file.name = "",
  ioe.file.name = ""
)
}
\arguments{
\item{das.analysis.result}{name of DAS analysis tools result file.}

\item{program}{name of DAS analysis tool. (one of rMATS, SUPPA, or spliceR)}

\item{pvalue.cutoff}{value of pvalue cutoff. default value is 0.05}

\item{dpsi.cutoff}{value of dPSI cutoff. default value is 0.1}

\item{gene.model}{gene model of reference. One of Refseq, Ensembl, or GENCODE. (spliceR only)}

\item{genome.version}{genome version of reference. One of hg18, GRCh19, or GRCh38. (spliceR only)}

\item{as.type}{AS event type. One of A3SS, A5SS, SE, MXE, or RI. (rMATS only)}

\item{gtf.file.name}{a GTF format file of reference. (SUPPA only)}

\item{ioe.file.name}{name of ioe file generated by SUPPA generateEvents command. (SUPPA)}
}
\value{
converter.result
}
\description{
Generate ASpedia-R input format from DAS analysis tools result.
}
\examples{
## rMATS
das.file.name <- “examples/rMATS_test.txt”
rmats.converter.result <- asr_converter(das.file.name, program=“rMATS”, as.type=“SE”)

##SUPPA
das.file.name <- “examples/SUPPA_test.txt”
gtf.file.name <- “examples/test_gtf.gtf”
ioe.file.name <- “examples/SUPPA_test.ioe”
suppa.converter.result <- asr_converter(das.file.name, program=“SUPPA”,
                                        gtf.file=gtf.file.name, ioe.file=ioe.file.name)

##spliceR
das.file.name <- “examples/spliceR_test.txt”
splilcer.converter.result <- asr_converter(das.file.name, program=“spliceR”,
                                          gene.model=“Ensembl”, genome.version=“GRCh38”)
}
