% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_plot.R
\name{asr_plot}
\alias{asr_plot}
\title{visualization of annotation result}
\usage{
asr_plot(
  annotation.result = "",
  gtf.file.name = "",
  gene.model = "Ensembl",
  genome.version = "GRCh38",
  gene.name = "",
  as.id = "",
  heights.list = "",
  list.of.plot = "",
  result.dir = ""
)
}
\arguments{
\item{annotation.result}{asr_annotation function result.}

\item{gtf.file.name}{a GTF format file of reference.}

\item{gene.model}{gene model of reference. One of Refseq, Ensembl, or GENCODE.}

\item{genome.version}{genome version of reference. One of hg18, GRCh19, or GRCh38.}

\item{gene.name}{gene name to be visualization.}

\item{as.id}{list of AS ID to be visualization.}

\item{heights.list}{positive integer vectors for track heights include height of ideogram and gene track.}

\item{list.of.plot}{list of DNA or RNA feature to be visualization. choose from "conservation", "NMD", "repeats", "domain", "PTM", or "RBP".}

\item{result.dir}{directory where plots(.png files) are saved}
}
\description{
DNA and RNA features in result of asr_annotation function to be visualize.
}
\examples{
gtf.file.name <- system.file("extdata", "test_gtf.gtf", package="ASpediaR")
plot.result.dir <- paste0(system.file("extdata", package="ASpediaR"), "/plot_result")

##using gene name
asr_plot(annotation.result, gtf.file.name, gene.model="Ensembl", genome.version="GRCh38",
          gene.name="FGFR2", result.dir=plot.result.dir)

##using AS ID
asr_plot(annotation.result, gtf.file.name, gene.model="Ensembl", genome.version="GRCh38", 
          as.id="chr10:121520169:121519979:121518829:121518682:121517463:121517319", 
          result.dir=plot.result.dir)

##track lists and track heights are change
asr_plot(annotation.result, gtf.file.name, gene.model="Ensembl", genome.version="GRCh38", 
          as.id="chr10:121520169:121519979:121518829:121518682:121517463:121517319", 
          heights.list=c(1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1), list.of.plot=c("conservation", "domain", "PTM", "repeats", "RBP"), 
          result.dir=plot.result.dir)
}
