% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mining_gsea.R
\name{mining_gsea}
\alias{mining_gsea}
\title{gene enrichment test between annotation gene list and knowledge-based database gene list.}
\usage{
mining_gsea(annotation.gene.list, gsea.gene.list, result.dir)
}
\arguments{
\item{annotation.gene.list}{gene list from asr_annotation function result}

\item{gsea.gene.list}{gene list from reference}

\item{result.dir}{directory where GSEA result(.tsv and .png file) are saved}
}
\description{
Gene enrichment test result are provided to tsv and plot format between annotation gene list and knowledge-based database gene list.
}
\examples{
## reference gene list from GTF
library(rtracklayer)
gtf.file.name <- system.file(“extdata”, “test_gtf.gtf”, package=“ASpediaR”)
gtf.data <- import(gtf.file.name)
gtf.gene.list <- unique(gtf.data$gene_name)
annotation.gene.list <- unique(annotation.result$gene_symbol)
gsea.result.dir <- system.file(“extdata/gsea_result”, package=“ASpediaR”)
mining_gsea(annotation.gene.list, gsea.gene.list=reference.gene.list,
              result.dir=gsea.result.dir)

## reference gene list from user input
test.gene.list.file.name <- system.file(“extdata”, “test_whole_gene.txt”, package=“ASpediaR”)
test.gene.list <- read.table(test.gene.list.file.name, header=FALSE, stringsAsFactors=FALSE)
gsea.result.dir <- system.file(“extdata/gsea_result”, package=“ASpediaR”)
mining_gsea(annotation.gene.list, gsea.gene.list=test.gene.list, result.dir=gsea.result.dir)
}
