% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceR_converter.R
\name{spliceR.converter}
\alias{spliceR.converter}
\title{spliceR result convert to ASpedia-R input format.}
\usage{
spliceR.converter(spliceR.result, pvalue.cutoff, gene.model, genome.version)
}
\arguments{
\item{spliceR.result}{name of spliceR result file.}

\item{pvalue.cutoff}{value of pvalue cutoff. default value is 0.05}

\item{gene.model}{gene model of reference. One of Refseq, Ensembl, or GENCODE.}

\item{genome.version}{genome version of reference. One of hg18, GRCh19, or GRCh38.}
}
\value{
converting.result
}
\description{
spliceR result convert to ASpedia-R input format.
}
\examples{
spliceR.result.file <- system.file(“extdata”, “spliceR_test.txt”, package=“ASpediaR”)
splilcer.converter.result <- spliceR_converter(spliceR.result.file, program=“spliceR”,
                                         gene.model=“Ensembl”, genome.version=“GRCh38”)
}
